Update codec usage
